requirements.txt
streamlit
streamlit_chat
langchain
python-dotenv
openai
tiktoken
langchain_openai